use Modal components to build out modal
